pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fussel"
version = "1.0.0"
description = "A static photo gallery generator"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Fussel Contributors"}
]
keywords = ["gallery", "photos", "static", "generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Pillow>=10.4.0",
    "beautifulsoup4>=4.12.3",
    "massedit>=0.70.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "jinja2-cli",
]

[project.scripts]
fussel = "fussel.fussel:main"

[tool.setuptools]
packages = ["fussel", "fussel.generator"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[project.optional-dependencies]
test = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "pytest-mock>=3.0.0"]
